+++ mne/bem.py
@@ -145,6 +145,7 @@ def _correct_auto_elements(surf, mat):
     for j, miss in enumerate(misses):
         # How much is missing?
         n_memb = len(surf['neighbor_tri'][j])
+        assert n_memb > 0  # should be guaranteed by our surface checks
         # The node itself receives one half
         mat[j, j] = miss / 2.0
         # The rest is divided evenly among the member nodes...
@@ -171,8 +172,8 @@ def _fwd_bem_lin_pot_coeff(surfs):
         rr_ord = np.arange(nps[si_1])
         for si_2, surf2 in enumerate(surfs):
             logger.info("        %s (%d) -> %s (%d) ..." %
-                        (_bem_explain_surface(surf1['id']), nps[si_1],
-                         _bem_explain_surface(surf2['id']), nps[si_2]))
+                        (_surf_name[surf1['id']], nps[si_1],
+                         _surf_name[surf2['id']], nps[si_2]))
             tri_rr = surf2['rr'][surf2['tris']]
             tri_nn = surf2['tri_nn']
             tri_area = surf2['tri_area']
@@ -252,11 +253,24 @@ def _fwd_bem_ip_modify_solution(solution, ip_solution, ip_mult, n_tri):
     return
 
 
+def _check_complete_surface(surf, copy=False):
+    surf = complete_surface_info(surf, copy=copy, verbose=False)
+    fewer = np.where([len(t) < 3 for t in surf['neighbor_tri']])[0]
+    if len(fewer) > 0:
+        raise RuntimeError('Surface %s has topological defects: %d / %d '
+                           'vertices have fewer than three neighboring '
+                           'triangles [%s]'
+                           % (_surf_name[surf['id']],
+                              len(fewer), surf['ntri'],
+                              ', '.join(str(f) for f in fewer)))
+    return surf
+
+
 def _fwd_bem_linear_collocation_solution(m):
     """Compute the linear collocation potential solution."""
     # first, add surface geometries
     for surf in m['surfs']:
-        complete_surface_info(surf, copy=False, verbose=False)
+        _check_complete_surface(surf)
 
     logger.info('Computing the linear collocation solution...')
     logger.info('    Matrix coefficients...')
@@ -483,7 +497,9 @@ def _surfaces_to_bem(surfs, ids, sigmas, ico=None, rescale=True,
         for si, surf in enumerate(surfs):
             surfs[si] = _ico_downsample(surf, ico)
     for surf, id_ in zip(surfs, ids):
+        # Do topology checks (but don't save data) to fail early
         surf['id'] = id_
+        _check_complete_surface(surf, copy=True)
         surf['coord_frame'] = surf.get('coord_frame', FIFF.FIFFV_COORD_MRI)
         surf.update(np=len(surf['rr']), ntri=len(surf['tris']))
         if rescale:
@@ -1233,7 +1249,7 @@ def read_bem_surfaces(fname, patch_stats=False, s_id=None, verbose=None):
             logger.info('    %d BEM surfaces read' % len(surf))
         for this in surf:
             if patch_stats or this['nn'] is None:
-                complete_surface_info(this, copy=False)
+                _check_complete_surface(this)
     return surf[0] if s_id is not None else surf
 
 
@@ -1424,18 +1440,12 @@ def _bem_find_surface(bem, id_):
         name = id_
         id_ = _surf_dict[id_]
     else:
-        name = _bem_explain_surface(id_)
+        name = _surf_name[id_]
     idx = np.where(np.array([s['id'] for s in bem['surfs']]) == id_)[0]
     if len(idx) != 1:
         raise RuntimeError('BEM model does not have the %s triangulation'
                            % name.replace('_', ' '))
     return bem['surfs'][idx[0]]
-
-
-def _bem_explain_surface(id_):
-    """Return a string corresponding to the given surface ID."""
-    _rev_dict = {val: key for key, val in _surf_dict.items()}
-    return _rev_dict[id_]
 
 
 # ############################################################################